Fixed menu problems.
1. Add animation for the menu transition (opacity 0->1). Note: to
change css sequentially, try to use .queue() function, and don’t forget
to add .dequeue() at the end.
2. There was an issue that when menu-icon is not at the top, click it,
then the page auto scroll, but one would wait a long time to see the
menu show up. The fix is simply add an if condition within hideMenu()
function since this function would be called multiple times, which
takes too much time.
3. The last problem is, the webpage has a wired white space on right.
To fix this, I found the trouble maker, which is the .meta class. The
way to find it is:
$(*).each(function(){
  rightPos = $(this).offset().left + $(this).width();
  // check the right position of every DOM element
 })